runtime: work around escaping closure in export_test.go
When compiling runtime, it is not allowed for local variables
and closures to be heap allocated. In one test, there is a go
statement with a closure. In the gc compiler, it distinguishes
capturing variable by value vs. by address, and rewrites it to
passing the captured values as arguments. Currently we don't
have this, and the escape analysis decides to heap allocate the
closure and also the captured variables, which is not allowed.
Work around it by passing the variables explicitly.

This is in preparation of turning on escape analysis for the
runtime.

Change-Id: Ie64e652afa6f07ea17be32deda433f5e92f6aae9
Reviewed-on: https://go-review.googlesource.com/86245
Reviewed-by: Ian Lance Taylor <iant@golang.org>