compiler: use temporary variable for stack allocation
Currently, allocation expression that can be allocated on stack
is implemented with __builtin_alloca, which turns into
__morestack_allocate_stack_space, which may call C malloc. This
may be slow. Also if this happens during certain runtime
functions (e.g. write barrier), bad things might happen (when
the escape analysis is enabled for the runtime). Make a temporary
variable on stack for the allocation instead.

Also remove the write barrier in the assignment in building heap
expression when it is stack allocated.

Change-Id: I71460c46d832825ee8bba296ac2a6680579a900c
Reviewed-on: https://go-review.googlesource.com/86242
Reviewed-by: Ian Lance Taylor <iant@golang.org>