go/types: implement SizesFor for gccgo
Move the architecture-specific settings out of configure.ac into a new
shell script goarch.sh.  Use the new script to collect the values for
all architectures to make them available in go/types.

Also fix cmd/vet to pass the right compiler when it calls SizesFor.

This fixes cmd/vet for systems that are not implemented in the gc
toolchain, such as alpha and ia64.

Change-Id: I177b4ee47a89402d0353644a4d87a8a6394cb0b2
Reviewed-on: https://go-review.googlesource.com/87635
Reviewed-by: Than McIntosh <thanm@google.com>