compiler: pass int to makechan, call makechan64 when appropriate
The update to 1.10beta1 changed makechan to take int instead of int64,
and added a makechan64 call for large values.  Since the size is the
last argument to makechan, the old compiler which always passed a
64-bit int worked fine on 64-bit systems and little-endian 32-bit
systems, but broke on big-endian 32-bit systems.  This CL fixes the
compiler to use the appropriate types.

This fixes GCC PR 83787.

Change-Id: I0f6da536293a8143d3d1faeab3eb16e05c3931cc
Reviewed-on: https://go-review.googlesource.com/88077
Reviewed-by: Than McIntosh <thanm@google.com>