libgo: break dependence on libgcc unwind-pe.h
The C portion of the Go runtime includes the header "unwind-pe.h" from
libgcc, which contains some constants and a few small routines for
decoding pointer values within unwind info. This patch gets rid of
that include and instead adds a re-implementation of that
functionality in the single file that uses it. The intent is to allow
the C runtime portion of libgo to be built without a companion GCC
installation.

Change-Id: I838185db423f0f2a9baa8752693f4d5de74eb340
Reviewed-on: https://go-review.googlesource.com/90235
Reviewed-by: Ian Lance Taylor <iant@golang.org>