net: rename TestAddr6 to avoid gotest confusion
On ppc64 gotest treats data variables whose names begin with "Test" as
tests to run.  This is to support the function descriptors used for
ppc64 ELF ABI v1.  This causes gotest to think that TestAddr6 is a
test, when it is actually a variable.  For a simple fix until we can
figure out how to write gotest properly, rename the variable.

Fixes golang/go#23623

Change-Id: Idc0398d0e0f622cbc6a4257b4fdfa805ff5008e9
Reviewed-on: https://go-review.googlesource.com/90995
Reviewed-by: Brad Fitzpatrick <bradfitz@golang.org>
Reviewed-by: Than McIntosh <thanm@google.com>