compiler: permit empty statements after fallthrough
The language spec permits empty statements after a fallthrough
statement, so implement that.  Also give a better error message when a
fallthrough statement is in the wrong place.  The test case for this
is in the master repository, test/fixedbugs/issue14540.go, just not
yet in the gccgo repository.

Fixes golang/go#14538

Change-Id: I1a8b409bb688e19727d66a8b46e22f8b4cacf746
Reviewed-on: https://go-review.googlesource.com/91855
Reviewed-by: Than McIntosh <thanm@google.com>