compiler: make single Btype for methods table of identical interface type
Normally we ensure to build a single Btype for identical types.
We did not do this for methods table of identical interface
types, however. If there are two identical interface type I, I2,
they have the same Btype BI, but different Btypes for their
methods tables, BM and BM2. From the backend's point of view
only one of them is linked to BI. This can cause inconsitency
in the backend's type system, like unresolved placeholder. This
CL ensures we create a single Btype for methods table of
identical interface type.

Change-Id: Ia464d9999af5564b190e0ec1d13d40cb719543d1
Reviewed-on: https://go-review.googlesource.com/92436
Reviewed-by: Ian Lance Taylor <iant@golang.org>