runtime: get missing function name from symbol table
If we trace back through code that has no debug info, as when calling
through C code compiled with -g0, we won't have a function name.
Try to fetch the function name using the symbol table.

Adding the test case revealed that gotest failed to use the gccgo tag
when matching files, so add that.

Change-Id: Id71facc637ff4b304fda9045ed1c251f6f3358e7
Reviewed-on: https://go-review.googlesource.com/92756
Reviewed-by: Than McIntosh <thanm@google.com>