compiler: error on func declaration/definition
Long long long ago Go permitted writing
    func F()
in one file and writing
    func F() {}
in another file.  This was removed from the language, and that is now
considered to be a multiple definition error.  Gccgo never caught up
to that, and it has been permitting this invalid code for some time.

Stop permitting it, so that we give correct errors.  Since we've
supported it for a long time, the compiler uses it in a couple of
cases: it predeclares the hash/equal methods if it decides to create
them while compiling another function, and it predeclares main.main as
a mechanism for getting the right warning if a program uses the wrong
signature for main.  For simplicity, keep those existing uses.

This required a few minor changes in libgo which were relying,
unnecessarily, on the current behavior.

Change-Id: Ie8835aad9775ec6417cad766f42336adde60dc79
Reviewed-on: https://go-review.googlesource.com/93083
Reviewed-by: Than McIntosh <thanm@google.com>