libgo: add type/const references to sysinfo.c
This patch adds explicit references to various types and constants
defined by the header files included by sysinfo.c (used to drive the
generation of gen-sysinfo.go as part of the libgo build via the GCC
"-fdump-go-spec" option).

The intent is to enable clients to gather the same info generated by
"-fdump-go-spec" by instead reading the generated DWARF from a
sysinfo.o object file compiled with "-g". Some compilers (notably
clang) try to omit DWARF records for a given type unless there is an
explicit use of it in the translation unit; the additional references
are to insure that everything we want to see in the DWARF shows up.

Change-Id: I154876888533fe03e4761176ec32c2a9345d0e36
Reviewed-on: https://go-review.googlesource.com/99063
Reviewed-by: Ian Lance Taylor <iant@golang.org>