Makefile: add internal/trace to noinst_DATA
The internal/trace package is only imported by tests (specifically the
tests in runtime/trace) so it must be in noinst_DATA to ensure that it
is built before running the tests.

This was mostly working because internal/trace has tests itself, and
is listed in check-packages.txt before runtime/trace, so typical
invocations of make would build internal/trace for checking purposes
before checking runtime/trace.  But we need this change to make that
reliable.

Change-Id: I315ec3f7b47f00e62d06dbc1ecfbb96ebeb0f631
Reviewed-on: https://go-review.googlesource.com/99836
Reviewed-by: Than McIntosh <thanm@google.com>