compiler: connect the concrete type and the placeholder for circular types
Previously, when creating the backend representation of a
circular type, we resolve the placeholder to a
circular_pointer_type. The backend doesn't know what the
concrete type is.

This CL changes it to resolve the placeholder to the concrete
type instead, so the backend may have better knowledge of the
concrete type.

Change-Id: I8fe458b71b6d4ac5f22d2f5e823c0af4f0320cb9
Reviewed-on: https://go-review.googlesource.com/123738
Reviewed-by: Ian Lance Taylor <iant@golang.org>