compiler: drop semicolons in export data
The export data, which is approximately readable and looks something
like Go, was first implemented back when Go still used semicolons.
Drop the semicolons, to make it look slightly more Go like and make it
slightly smaller.

This updates the compiler and the gccgoimporter package.

This introduces a new version of the export data.  There are going to
be more changes to the export data, so this version is still subject
to change.

Change-Id: Id2474c22a97c42501d81c38f732cf6c9b7f4ba21
Reviewed-on: https://go-review.googlesource.com/c/143018
Reviewed-by: Than McIntosh <thanm@google.com>