cmd/go: pass down testing gccgo in TestScript
This permits TestScript to work when gccgo is not installed.
Previous testing was using a previously installed gccgo, not the newly
built one.

This revealed that the testing of whether an internal package is
permitted was incorrect for standard library packages, since the
uninstalled gccgo can see internal packages in the uninstalled libgo.
Fix the internal package tests.

This permitted removing a couple of gccgo-specific changes in the
testsuite.

Change-Id: I8029cd73eba94c0b711f47af0df96f673489efc2
Reviewed-on: https://go-review.googlesource.com/137255
Reviewed-by: Than McIntosh <thanm@google.com>