runtime: skip testSetPanicOnFault for gollvm
LLVM doesn't support non-call exception. This test was passing
more or less by luck: if the faulting instruction is between two
calls with the same landing pad (in instruction layout order,
not the program's logic order), it generates a merged PC range
that covers the faulting instruction. If the instruction layout
order changes, or it uses two different (but may be degenerate)
landing pads, this doesn't work.

Change-Id: I92b85385476be0275016953a9f90b03c01ca66e2
Reviewed-on: https://go-review.googlesource.com/c/140517
Reviewed-by: Ian Lance Taylor <iant@golang.org>