compiler: use the underlying type to build placeholder type for alias
When asking for a placeholder type of an alias type, build a
placeholder for the underlying type, instead of treating the
alias as a named type and calling get_backend. The latter may
fail as we may not be ready to build a complete backend type. We
have already used a unified backend type for alias type and its
underlying type. Do the same for placeholders as well.

Change-Id: I4691b970f885ae60e1e2a27384c651dbba59feb7
Reviewed-on: https://go-review.googlesource.com/138635
Reviewed-by: Ian Lance Taylor <iant@golang.org>