sync/atomic: use strong form of atomic_compare_exchange_n
In the recent change to use atomic_compare_exchange_n I thought we
could use the weak form, which can spuriously fail. But that is not
how it is implemented in the gc library, and it is not what the rest
of the library expects.

Thanks to Lynn Boger for identifying the problem.

Fixes https://gcc.gnu.org/PR89199

Change-Id: I78dea34f1b92b84ec4367b4029a6253d599858fa
Reviewed-on: https://go-review.googlesource.com/c/161359
Reviewed-by: Cherry Zhang <cherryyz@google.com>